update: isolate menu and hook
